Re merge complete - moved to double value for NPeople and sim still works

+++ validated/src/main/java/uk/ac/standrews/cs/digitising_scotland/verisim/populationStatistics/dataDistributionTables/selfCorrecting/SelfCorrectingProportionalDistribution.java
@@ -72,7 +72,7 @@ public class SelfCorrectingProportionalDistribution implements DataDistribution 
                     new IntegerRangeToDoubleSet(Collections.singleton(new IntegerRange(1)), 0.0));
         }
         Integer sumOfAC = achievedCountsForAge.getSumOfValues();
-        Integer totalCount = sumOfAC + key.getForNPeople();
+        Double totalCount = sumOfAC + key.getForNPeople();
 
         LabeledValueSet<IntegerRange, Double> rawCorrectedValues =
                     targetProportions.get(resolveRowValue(age))
