+++ validated/src/main/java/uk/ac/standrews/cs/digitising_scotland/verisim/populationStatistics/dataDistributionTables/statsKeys/PartneringStatsKey.java
@@ -32,7 +32,7 @@ public class PartneringStatsKey extends StatsKey {
         return getYLabel();
     }
 
-    public int getForNumberOfFemales() {
+    public double getForNumberOfFemales() {
         return getForNPeople();
     }
 }
